REFACTOR: Refactored much of the selection handling code into its own class (SnapLinks.Selection) and some of the main code into a main class (SnapLinks).  This was probably already too much too commit into one revision so I decided to stop here and commit what I had.  Should be all in working order.  Mostly the code has just been moved around and many globals removed.  Much work remains to be done to refactor into really well maintained code.
Note: Added a very stripped down version of prototype's class creation functionality, it's really clean and easy to use, supports getters/setters.  I also have a snipped in there to allow for C# style setter/getter declarations in JS, makes for easily readable code.

I did comment out one block of code that would, in theory, be there for a 'feature' though I'm not sure of the utility of the feature, it is just commented out.

There were one or two other sections that appeared to have no effect, I commented those out and left a note as well.  Looks like I'm done for this iteration.