BUGFIX/REFACTOR: Fixes the issue of SLP breaking when a document is reloaded during a drag event.
When a document is refreshed during a drag event, the dead object issue arises.  This is a considerable refactoring of the way elements are stored and handled.  Most of the elements are now stored within the document being worked with.  If the document is reloaded our stored data vaporizes and through the use of getters most of them are automatically re-calculated.  In addition, when the top document is unloaded, mousemove/keyup/keydown events are suspended until the new top document is 'load'ed again.

This creates a few new issues which I will address shortly, the most of which is that the snap rect disappears and does not re-appear until the new top document is completely loaded.  If the user keeps the drag going and is patient, it will pick up where it left off, but this presents the user with an unusual state.  I would like to change that, but it will take some time to figure out a better/different way of dealing with the selection rect, divorcing it from the document its-self so that it can continue to be present and interactive while the document is re-loading.